source/backend/src/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0.1
pydantic>=2.6
langchain-openai>=0.3.0
langgraph>=0.2.0
requests>=2.31.0
